Move NotebookPage and extract helpers
