+++ src/pages/NotebookPage.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { TrpcProvider } from "@/components/TrpcProvider.tsx";
-import { NotebookPage as NotebookPageComponent } from "@/components/notebooks/NotebookPage.tsx";
+import { NotebookPage as NotebookPageComponent } from "@/components/notebooks/notebook/NotebookPage";
 
 export const NotebookPage: React.FC = () => {
   return (
